Year2024_day02: add solution.
I found that this recursive solution was pretty neat, pattern matching worked
really great here.

I think this could potentially be optimized by using array instead of lists, and
swapping elements around in the array to simulate skipping elements. I've done
something like that for Go and it resulted in some decent speedups. But at this
moment I'm not particularly interested in porting that solution here.