Add `ppx_sexp_conv` as a dependency and a playground file to exercise it.
I think I'm going to use this quite a bit for printing and stuff in the future.
I can use it without declaring an explicit package dependency on it, so I assume
it's in my dependency closure somewhere, but I haven't bothered to go looking
where. Besides, explicit dependencies are good.