Add solution for year 2024, day 05.
Probably not the most elegant way to do it, but I wanted to experiment a bit
with OCaml modules to understand more about how to define them and use them.

I haven't benchmarked this at all, can't be arsed. My intuition tells me that
maybe it'd be a good idea to put stuff in arrays instead of lists, since we're
dealing with sorting and stuff. And I used arrays anyway to get to the "middle"
element.