bin/benchmark: generate code for benchmarks.
This program reads the list of OCaml modules from the `lib`
folder; assumes they have functions called `part1` and
`part2`; assumes a corresponding input file exists; and generates the code for a
simple program that will run benchmarks for all solution functions.

It's a quick and dirty way of always having up-to-date benchmarks every time I
add a new solution.