lib/inputs: be less aggressive about stripping whitespace.
There are some problems (like year 2025, day 06) where the leading whitespace is
significant. For those the default behavior of `String.strip` significantly
changes the shape of the input. To work around that, only strip newlines
(primarily to strip trailing newlines, if any) and use the `{xxx|<string>|xxx}`
syntax to hopefully be less exposed to OCaml's ideas about whitespace in string
literals.