Year2025_day11: add solution.
This was a fun problem! I always tend to like the graph search ones. It annoys
me a little bit that it's so much ceremony around setting up a type for hash
keys in OCaml, where you have to declare a module and derive a bunch of
functions; coming from Go where the compiler can automatically make any
`comparable` type a hash key, this feels like a step back in convenience. Maybe
there is some ppx rewriter out there to make this easier.

Runs fast enough too on my laptop:

    │ Year2025_day11.part1 │     259.22us │     40.95kw │     15.83kw │     14.29kw │      0.05% │
    │ Year2025_day11.part2 │     739.08us │     92.99kw │     35.18kw │     30.06kw │      0.13% │